refactor: Socket exam restructure
- refactor CRUD logic related to exam creation and data validation to new controller factory function `createExam` from `handleExams.js` to a new module `examControllers.js`.
- add additional validation and support for factory models and error handling in `createExam` in `examControllers.js`.
- change handler function from `createExam` to `handleCreateExam` in `handleExams.js` and update the instances of handler name reference in  `manageExams.js`.
